backend/requirements.txt
Django>=5.0,<5.2
daphne>=4.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24.2
psycopg[binary]>=3.2
channels>=4.1
channels-redis>=4.2
celery[redis]>=5.4
redis>=5.0
gunicorn>=22.0
whitenoise>=6.7
python-dotenv>=1.0
